+++ src/Products/urban/__init__.py
@@ -151,10 +151,14 @@ def initialize(context):
     import Locality
     import OpinionRequestEventType
     import OrganisationTerm
+    # [DX] TO DELETE > 2.5 [DX]
     import ParcellingTerm
+    # [DX] TO DELETE > 2.5 [DX]
     import PcaTerm
     import PersonTitleTerm
+    # [DX] TO DELETE > 2.5 [DX]
     import PortionOut
+    # [DX] TO DELETE > 2.5 [DX]
     import Recipient
     import RecipientCadastre
     import SpecificFeatureTerm