+++ src/Products/urban/browser/table/tablevalue.py
@@ -188,14 +188,6 @@ class ValuesForClaimantsListing(ValuesForUrbanListing):
         return claimants
 
 
-class ValuesForFolderListing(ValuesForUrbanListing):
-    """  return values from the context """
-
-    def getItems(self):
-        items = self.context.objectValues()
-        return items
-
-
 class ValuesForLicenceListing(ValuesForUrbanListing):
     """ return licence values from the context  """