+++ docs/CHANGES.rst
@@ -29,6 +29,39 @@ Changes
 - Fix default schedule config for CODT Buildlicence [SUP-12344]
   [sdelcourt]
 
+- Allow shortcut transition to 'inacceptable' state for CODT licence wofklow. [SUP-6385]
+  [sdelcourt]
+
+- Set default foldermanagers view to sort the folder with z3c.table on title [URB-1151]
+  [jjaumotte]
+
+- Add some applicants infos on urban_description schemata. [URB-1171]
+  [jjaumotte]
+
+- Improve default reference expression for licence references. [URB-2046]
+  [sdelcourt]
+
+- Add search filter on public config folders (geometricians, notaries, architects, parcellings). [SUP-10537]
+  [sdelcourt]
+
+- Migrate PortionOut (Archetype) type to Parcel (dexterity) type. [URB-2009]
+  [sdelcourt]
+
+- Fix add permissions for Inquiries. [SUP-13679]
+  [sdelcourt]
+
+- Add custom division 99999 for unreferenced parcels. [SUP-13835]
+  [sdelcourt]
+
+- Migrate ParcellingTerm (Archetype) type to Parcelling (dexterity) type.
+  [sdelcourt]
+
+- Add field to define final states closing all the urban events on a licence. [URB-2082]
+  [sdelcourt]
+
+- Refactor key date display to include urban event custom titles. [SUP-13982]
+  [sdelcourt]
+
 
 2.4 (2019-03-25)
 ----------------
@@ -101,8 +134,12 @@ Changes
 - Add migration source code
   [fngaha]
 
+2.3.x (unreleased)
+-------------------
+- Update MultipleContactCSV methods with an optional number_street_inverted (#17811)
+  [jjaumotte]
 
-1.11.1 (unreleased)
+1.11.1 (unknown release date)
 -------------------
 - add query_parcels_in_radius method to view
   [fngaha]